GDI-38 | Review comments addressed
