+++ src/main/java/com/credenceid/issuance/service/controller/IssuanceController.java
@@ -1,9 +1,11 @@
 package com.credenceid.issuance.service.controller;
 
-import com.credenceid.issuance.service.services.IssuanceService;
+import com.credenceid.identity.iso18013.cosekey.CoseKey;
+import com.credenceid.issuance.service.domain.model.Application;
+import com.credenceid.issuance.service.domain.model.DigitalIdResult;
+import com.credenceid.issuance.service.domain.model.PersonalData;
 import com.credenceid.issuance.service.vo.ApplicationRequest;
 import com.credenceid.issuance.service.vo.DigitalIdResponse;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,14 +13,14 @@ import org.springframework.web.bind.annotation.RestController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.nio.charset.StandardCharsets;
+
 @RequestMapping(value = "/issuanceservice")
 @RestController
 public class IssuanceController {
 
     private static final Logger logger = LoggerFactory.getLogger(IssuanceController.class);
 
-    @Autowired
-    IssuanceService service;
 
     /**
      * Endpoint to generate digital id from application data
@@ -31,8 +33,46 @@ public class IssuanceController {
 
         logger.info("WalletToken : {}, DeviceKey : {}, Personal Data : {}",
                 applicationRequest.getWalletToken(), applicationRequest.getDeviceKey(), applicationRequest.getPersonalData());
+        Application application = getApplicationData(applicationRequest);
 
-        return service.generateDigitalId(applicationRequest);
+        /* created dummy response */
+        //TODO implement generation of digitalId
+        DigitalIdResult result = new DigitalIdResult(application.walletToken(), "testDigitalKey".getBytes(StandardCharsets.UTF_8), "testNameSpaces".getBytes(StandardCharsets.UTF_8));
+        DigitalIdResponse response = new DigitalIdResponse();
+        response.setWalletToken(result.walletToken());
+        response.setMso(result.mso());
+        response.setNameSpaces(result.nameSpaces());
+        return response;
+    }
+
+    /**
+     * fetch PersonalData from applicationRequest
+     *
+     * @param applicationRequest
+     * @return PersonalData
+     */
+    private PersonalData getPersonalData(ApplicationRequest applicationRequest) {
+        return new PersonalData(
+                applicationRequest.getPersonalData().getPortrait(),
+                applicationRequest.getPersonalData().getDocumentCode(),
+                applicationRequest.getPersonalData().getIssuer(),
+                applicationRequest.getPersonalData().getDocumentNumber(),
+                applicationRequest.getPersonalData().getDateOfBirth(),
+                applicationRequest.getPersonalData().getSex(),
+                applicationRequest.getPersonalData().getDateOfExpiry(),
+                applicationRequest.getPersonalData().getFamilyName(),
+                applicationRequest.getPersonalData().getGivenNames());
+    }
+
+    /**
+     * fetch ApplicationData from ApplicationRequest
+     *
+     * @param applicationRequest
+     * @return Application
+     */
+    private Application getApplicationData(ApplicationRequest applicationRequest) {
+        CoseKey deviceKey = new CoseKey(applicationRequest.getDeviceKey());
+        return new Application(getPersonalData(applicationRequest), deviceKey, applicationRequest.getWalletToken());
     }
 
 }
