--- src/main/java/com/credenceid/issuance/service/services/IssuanceService.java
@@ -1,8 +1,0 @@
-package com.credenceid.issuance.service.services;
-
-import com.credenceid.issuance.service.vo.ApplicationRequest;
-import com.credenceid.issuance.service.vo.DigitalIdResponse;
-
-public interface IssuanceService {
-    public DigitalIdResponse generateDigitalId(ApplicationRequest req);
-}
